+++ services/web/mgmt/2020-06-01/web/models.go
@@ -2262,7 +2262,7 @@ type AppServiceCertificateOrderPatchResourceProperties struct {
 	// IsPrivateKeyExternal - READ-ONLY; <code>true</code> if private key is external; otherwise, <code>false</code>.
 	IsPrivateKeyExternal *bool `json:"isPrivateKeyExternal,omitempty"`
 	// AppServiceCertificateNotRenewableReasons - READ-ONLY; Reasons why App Service Certificate is not renewable at the current moment.
-	AppServiceCertificateNotRenewableReasons *[]string `json:"appServiceCertificateNotRenewableReasons,omitempty"`
+	AppServiceCertificateNotRenewableReasons *[]ResourceNotRenewableReason `json:"appServiceCertificateNotRenewableReasons,omitempty"`
 	// NextAutoRenewalTimeStamp - READ-ONLY; Time stamp when the certificate would be auto renewed next
 	NextAutoRenewalTimeStamp *date.Time `json:"nextAutoRenewalTimeStamp,omitempty"`
 }
@@ -2331,7 +2331,7 @@ type AppServiceCertificateOrderProperties struct {
 	// IsPrivateKeyExternal - READ-ONLY; <code>true</code> if private key is external; otherwise, <code>false</code>.
 	IsPrivateKeyExternal *bool `json:"isPrivateKeyExternal,omitempty"`
 	// AppServiceCertificateNotRenewableReasons - READ-ONLY; Reasons why App Service Certificate is not renewable at the current moment.
-	AppServiceCertificateNotRenewableReasons *[]string `json:"appServiceCertificateNotRenewableReasons,omitempty"`
+	AppServiceCertificateNotRenewableReasons *[]ResourceNotRenewableReason `json:"appServiceCertificateNotRenewableReasons,omitempty"`
 	// NextAutoRenewalTimeStamp - READ-ONLY; Time stamp when the certificate would be auto renewed next
 	NextAutoRenewalTimeStamp *date.Time `json:"nextAutoRenewalTimeStamp,omitempty"`
 }
@@ -21697,6 +21697,10 @@ type SiteConfig struct {
 	RemoteDebuggingVersion *string `json:"remoteDebuggingVersion,omitempty"`
 	// HTTPLoggingEnabled - <code>true</code> if HTTP logging is enabled; otherwise, <code>false</code>.
 	HTTPLoggingEnabled *bool `json:"httpLoggingEnabled,omitempty"`
+	// AcrUseManagedIdentityCreds - Flag to use Managed Identity Creds for ACR pull
+	AcrUseManagedIdentityCreds *bool `json:"acrUseManagedIdentityCreds,omitempty"`
+	// AcrUserManagedIdentityID - If using user managed identity, the user managed identity ClientId
+	AcrUserManagedIdentityID *string `json:"acrUserManagedIdentityID,omitempty"`
 	// LogsDirectorySizeLimit - HTTP logs directory size limit.
 	LogsDirectorySizeLimit *int32 `json:"logsDirectorySizeLimit,omitempty"`
 	// DetailedErrorLoggingEnabled - <code>true</code> if detailed error logging is enabled; otherwise, <code>false</code>.
@@ -21832,6 +21836,12 @@ func (sc SiteConfig) MarshalJSON() ([]byte, error) {
 	}
 	if sc.HTTPLoggingEnabled != nil {
 		objectMap["httpLoggingEnabled"] = sc.HTTPLoggingEnabled
+	}
+	if sc.AcrUseManagedIdentityCreds != nil {
+		objectMap["acrUseManagedIdentityCreds"] = sc.AcrUseManagedIdentityCreds
+	}
+	if sc.AcrUserManagedIdentityID != nil {
+		objectMap["acrUserManagedIdentityID"] = sc.AcrUserManagedIdentityID
 	}
 	if sc.LogsDirectorySizeLimit != nil {
 		objectMap["logsDirectorySizeLimit"] = sc.LogsDirectorySizeLimit